Updated rust-builder to somthing post-async/await
Substrate stopped building a docker image. We need build it ourselves now.

The container still doesn't run properly, but that has nothing to to with docker.

Apparently, the current `substrate --version` (substrate 2.0.0-aa937d9b4) assumes
the wasm runtime will export "BabeApi_configuration". During out recent dependency
update (0617864ff6), the babe was replaced with aura to match the current
node-template. I will switch the runtime back to babe, but not on this branch.